% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdPop.R
\name{sd.population}
\alias{sd.population}
\title{Calculate standard deviation from a population instead of a sample}
\usage{
sd.population(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{Logical. Remove missing values.}
}
\description{
This function calculates the standard deviation from a population (this means it
is divided by \emph{n}, not by \emph{n-1}). It returns a numeric value indicating the standard
deviation for a given numeric vector.
}
\examples{
x <- runif(50)
sd.population(x)
}
\keyword{deviation,}
\keyword{population}
\keyword{standard}
